+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # :arrows_clockwise: Dell EMC VxRail As Built Report Changelog
 
+## [[0.3.1](https://github.com/AsBuiltReport/AsBuiltReport.DellEMC.VxRail/releases/tag/v0.3.1)] - 2022-07-08
+### Added
+- Added sample Dell EMC VxRail As Built Report
+### Fixed
+- Fixed GitHub Release workflow
+### Changed
+- Improved report formatting
+
 ## [[0.3.0](https://github.com/AsBuiltReport/AsBuiltReport.DellEMC.VxRail/releases/tag/v0.3.0)] - 2021-09-10
 ### Added
 * PowerShell 7 compatibility